InitialCommit
While not all tests have been written or included, this is functional
for a well-written bit-code file.  Work that remains is converting from
crashing to throwing errors for non-well-formed files, and create
sample data for better and complete testing, though complete end-to-end
has been tested on real bitcode files not included here.